Disable the local part of package versions (#10)
PyPI does not accept Python packages with a local part in the version
number (e.g. +g84953ff). When building the packages for a commit on the
main branch, the local part is now turned off by temporarily modifying
the setuptools_scm settings in pyproject.toml.

Packages build form other branches retain the local part so as to not
have conflicting version numbers with those build from the main
branch.

As a consequence all packages, also the .dev versions, build from the
main branch can now be published once again on PyPI.

(DIS-1584)